feat: releases overview shows total docs in release for open releases
